User - Comment : Delete

+++ articles/views.py
@@ -77,7 +77,12 @@ def comment_create(request, pk):
         comment.save()
     return redirect('articles:detail', article.pk)
 
+@login_required
 def comments_delete(request, article_pk, comment_pk):
     comment = Comment.objects.get(pk=comment_pk)
-    comment.delete()
-    return redirect('articles:detail', article_pk)
+    if request.user == comment.user:
+        if request.method == 'POST':
+            comment.delete()
+            return redirect('articles:detail', article_pk)
+    else:
+        return HttpResponseForbidden()
